refactor FundWalletContent
